refactored folder structure to prepare for new admin section
