--- src/components/Home.js
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-import CheckIn from './CheckIn'
-
-export default class Home extends Component {
-  render() {
-    return (
-      <CheckIn />
-    )
-  }
-}
