+++ src/components/auth/Login.js
@@ -1,35 +1,47 @@
 import React, { Component } from 'react';
-import { auth } from '../helpers/auth';
+import { login, resetPassword } from '../../helpers/auth';
 
 import RaisedButton from 'material-ui/RaisedButton';
 import TextField from 'material-ui/TextField';
 
 function setErrorMsg(error) {
   return {
-    registerError: error.message
+    loginMessage: error
   };
 }
 
-export default class Register extends Component {
+export default class Login extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      registerError: null,
       email: '',
-      password: ''
+      password: '',
+      loginMessage: null
     };
   }
 
   handleSubmit = e => {
     e.preventDefault();
-    auth(this.state.email, this.state.password).catch(e =>
-      this.setState(setErrorMsg(e))
-    );
+    login(this.state.email, this.state.password).catch(error => {
+      this.setState(setErrorMsg('Invalid username/password.'));
+    });
+  };
+  resetPassword = () => {
+    resetPassword(this.state.email)
+      .then(() =>
+        this.setState(
+          setErrorMsg(`Password reset email sent to ${this.state.email}.`)
+        )
+      )
+      .catch(error => this.setState(setErrorMsg(`Email address not found.`)));
   };
   render() {
     return (
-      <form onSubmit={this.handleSubmit} style={style.container}>
-        <h3>Register</h3>
+      <form
+        style={style.container}
+        onSubmit={event => this.handleSubmit(event)}
+      >
+        <h3>Login</h3>
         <TextField
           hintText="Enter your Email"
           floatingLabelText="Email"
@@ -43,18 +55,21 @@ export default class Register extends Component {
           onChange={(event, newValue) => this.setState({ password: newValue })}
         />
         <br />
-        {this.state.registerError && (
+        {this.state.loginMessage && (
           <div className="alert alert-danger" role="alert">
             <span
               className="glyphicon glyphicon-exclamation-sign"
               aria-hidden="true"
             />
             <span className="sr-only">Error:</span>
-            &nbsp;{this.state.registerError}
+            &nbsp;{this.state.loginMessage}{' '}
+            <p onClick={this.resetPassword} className="alert-link">
+              Forgot Password?
+            </p>
           </div>
         )}
         <RaisedButton
-          label="Register"
+          label="Login"
           primary={true}
           style={style.raisedBtn}
           type="submit"
