+++ src/components/auth/Register.js
@@ -1,47 +1,35 @@
 import React, { Component } from 'react';
-import { login, resetPassword } from '../helpers/auth';
+import { auth } from '../../helpers/auth';
 
 import RaisedButton from 'material-ui/RaisedButton';
 import TextField from 'material-ui/TextField';
 
 function setErrorMsg(error) {
   return {
-    loginMessage: error
+    registerError: error.message
   };
 }
 
-export default class Login extends Component {
+export default class Register extends Component {
   constructor(props) {
     super(props);
     this.state = {
+      registerError: null,
       email: '',
-      password: '',
-      loginMessage: null
+      password: ''
     };
   }
 
   handleSubmit = e => {
     e.preventDefault();
-    login(this.state.email, this.state.password).catch(error => {
-      this.setState(setErrorMsg('Invalid username/password.'));
-    });
-  };
-  resetPassword = () => {
-    resetPassword(this.state.email)
-      .then(() =>
-        this.setState(
-          setErrorMsg(`Password reset email sent to ${this.state.email}.`)
-        )
-      )
-      .catch(error => this.setState(setErrorMsg(`Email address not found.`)));
+    auth(this.state.email, this.state.password).catch(e =>
+      this.setState(setErrorMsg(e))
+    );
   };
   render() {
     return (
-      <form
-        style={style.container}
-        onSubmit={event => this.handleSubmit(event)}
-      >
-        <h3>Login</h3>
+      <form onSubmit={this.handleSubmit} style={style.container}>
+        <h3>Register</h3>
         <TextField
           hintText="Enter your Email"
           floatingLabelText="Email"
@@ -55,21 +43,18 @@ export default class Login extends Component {
           onChange={(event, newValue) => this.setState({ password: newValue })}
         />
         <br />
-        {this.state.loginMessage && (
+        {this.state.registerError && (
           <div className="alert alert-danger" role="alert">
             <span
               className="glyphicon glyphicon-exclamation-sign"
               aria-hidden="true"
             />
             <span className="sr-only">Error:</span>
-            &nbsp;{this.state.loginMessage}{' '}
-            <p onClick={this.resetPassword} className="alert-link">
-              Forgot Password?
-            </p>
+            &nbsp;{this.state.registerError}
           </div>
         )}
         <RaisedButton
-          label="Login"
+          label="Register"
           primary={true}
           style={style.raisedBtn}
           type="submit"
