+++ src/components/customer/index.js
@@ -1,0 +1,5 @@
+export {default as BottomNavigationBar } from './BottomBar';
+export {default as CheckIn} from './CheckIn';
+export {default as OpenTabs} from './OpenTabs';
+export {default as Tab} from './Tab';
+export {default as TabItems} from './TabItems';
