--- src/components/dummyVenues.js
@@ -1,440 +1,0 @@
-
-//function for how to loop through data
- // let venues = dummyVenues.results.map(venue => ({
-    //   merchant: venue.name,
-    //   address: venue.vicinity,
-    //   types: venue.types || [],
-    //   open: venue.opening_hours 
-    //     ? venue.opening_hours.open_now 
-    //     : false,
-    // }));
-    // const openRestaurants = venues.filter(venue => {
-    //   return venue.types.includes("restaurant") && venue.open
-    // })
-
-module.exports = {
-  "html_attributions": [],
-  "results": [
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.7050604,
-                  "lng": -74.00865979999999
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.7064093802915,
-                      "lng": -74.00731081970848
-                  },
-                  "southwest": {
-                      "lat": 40.7037114197085,
-                      "lng": -74.0100087802915
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/bar-71.png",
-          "id": "a2f2678d103a243c02b092c2099ee0a59e940e04",
-          "name": "Killarney Rose",
-          "opening_hours": {
-              "open_now": true,
-              "weekday_text": []
-          },
-          "photos": [
-              {
-                  "height": 3120,
-                  "html_attributions": [
-                      "<a href=\"https://maps.google.com/maps/contrib/106481933208769545164/photos\">Dave Beever</a>"
-                  ],
-                  "photo_reference": "CmRaAAAAUkHRXuCtnpKCEvsfCijibMGtoYNNabN-caJMnFWwp9hSHMxsgDoUIN7_a0B_9HINqJVpHKgvMOTy7-9yRy_6XLl2umiXWOEeIroNhimTcUTGJiHTAf1nCqqdYdt2v2E9EhAWgylrLgIxLkArxccCwwR5GhSD1cshHi7OIh2VzG3SIkSrbbhULw",
-                  "width": 4160
-              }
-          ],
-          "place_id": "ChIJQalvcBZawokRGSuWUh6cZiI",
-          "price_level": 2,
-          "rating": 4.1,
-          "reference": "CmRRAAAANaSioKSAShM8BT2BF2vKtwZlMOaLiB3h2sztpv_I0dqJUTDEBymqPIXpx_kbFVqfUdYjWoQod1bWq_aB2az6Vf7MEIilPL80qsnIhRL85NjPvJf-SYFWEO8_SSbQY9g1EhBcVnejteJYP1rJtJceJv_WGhRBio87MT9YNkbSNhwfZRpHmlholw",
-          "scope": "GOOGLE",
-          "types": [
-              "bar",
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "127 Pearl St, New York"
-      },
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.70523190000001,
-                  "lng": -74.00842999999999
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.7066693802915,
-                      "lng": -74.00691446970849
-                  },
-                  "southwest": {
-                      "lat": 40.7039714197085,
-                      "lng": -74.0096124302915
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/restaurant-71.png",
-          "id": "1a3facc89ce6203c6192dcb92433cc145a52d07b",
-          "name": "Haru Sushi",
-          "opening_hours": {
-              "open_now": true,
-              "weekday_text": []
-          },
-          "photos": [
-              {
-                  "height": 900,
-                  "html_attributions": [
-                      "<a href=\"https://maps.google.com/maps/contrib/115617818796129632614/photos\">Haru Sushi</a>"
-                  ],
-                  "photo_reference": "CmRaAAAAxeu4ZRmIiBPuxUqXwRQ26_yUw-y8ICgiFT4yVbjtgcO4PMccDhJYGFo8Z1Mbaw-Vg-dd3xal33AHxJmGV5piMcucV0r8_AxHeKXoG8VfF2YGPCebFWmm_ee7FMYmm1JPEhBFgx8PJUFsLWMqdgkqNMm_GhQ7jDLllqOzCRVg4h4yVmdO2q1hRw",
-                  "width": 1200
-              }
-          ],
-          "place_id": "ChIJe561bxZawokR7iWnSvqBJRE",
-          "price_level": 2,
-          "rating": 4.1,
-          "reference": "CmRRAAAAIIH8KjpRe7Xl-ORMeetrWXVYTPvMm9MDH6lOApQa9zUCCEorxDh7l6V9ks-d95z3g-M2C3MUU5HNwardykAab3VsW_n4xbbUgbS3IL-Bs-cCrlALVhH8I7jTOBVdHiIXEhDu6x9N-EiK_uguDdlpcbmNGhSKsFRfGpmc-32QNydRL2BIjrsC5g",
-          "scope": "GOOGLE",
-          "types": [
-              "bar",
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "1 Wall Street, New York"
-      },
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.7053349,
-                  "lng": -74.00836
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.70668163029149,
-                      "lng": -74.00701066970849
-                  },
-                  "southwest": {
-                      "lat": 40.70398366970849,
-                      "lng": -74.0097086302915
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/restaurant-71.png",
-          "id": "505a1d630b1dd16c2c999668908f210a642ec83a",
-          "name": "Fino",
-          "opening_hours": {
-              "open_now": false,
-              "weekday_text": []
-          },
-          "photos": [
-              {
-                  "height": 2048,
-                  "html_attributions": [
-                      "<a href=\"https://maps.google.com/maps/contrib/104430269876095674561/photos\">Ward Blackwell</a>"
-                  ],
-                  "photo_reference": "CmRaAAAAJqZe-UZLtBEZoxJZY_5cT6GXgHojdQX-JWBK7z-oRHEKWLMAPQDho7JoTheJvacl6fxh_XfmIc360NvL7JYjbhx5LAU8ubQXMdEgbWj8RBZpH9sXwjVhrxa9PQ0qPB5bEhBq0eC24-cev3UBgcNttdWxGhSwI9OvumXY4liQIi2pm0E4iriXcQ",
-                  "width": 1152
-              }
-          ],
-          "place_id": "ChIJ8VK4bxZawokRO0cBg-JYRBU",
-          "price_level": 2,
-          "rating": 4.2,
-          "reference": "CmRRAAAAq-OrUDJW7i8YDAk1OEUUJD8qikH3qCWHlcGWREXxYyXv-xxWwTdFqb0tG1dMMCkr7oAOSxbEDUIkIOlG5rsgG2TgGMUICd498TBJDoG5nJUfjl3Fpvdm6civEE_r5Jo0EhBTtuYgwPGUW5zgzr1ASVsCGhSvdPHy6CvcgPgGQCxxc_ik5iblog",
-          "scope": "GOOGLE",
-          "types": [
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "82 Beaver Street, New York"
-      },
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.70516050000001,
-                  "lng": -74.0087442
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.7065914802915,
-                      "lng": -74.00740926970848
-                  },
-                  "southwest": {
-                      "lat": 40.7038935197085,
-                      "lng": -74.0101072302915
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/generic_business-71.png",
-          "id": "b8d49b37973e677f20e48bdf6246ce99ef9a2f05",
-          "name": "80 Enterprises Inc. (Killarney Rose)",
-          "opening_hours": {
-              "open_now": true,
-              "weekday_text": []
-          },
-          "place_id": "ChIJHYh_ZRZawokR3tYN_Psp470",
-          "reference": "CmRSAAAA8xxHcvgEbwprMXkJtHfZgBfbkyekIeddje76XHBeCM9DT67SxFaYNZxOTLwvsw49KlUWe9EZMEaOi3REEKqnEjX2zKYR74VchEBrIGo0z_MO3kLAtq-5I0hJpwh7PbNqEhAw1_Nrew8sXlfk7VP9n9WuGhQ5alj--dnyQtbVpyqNo0j-FgwdNw",
-          "scope": "GOOGLE",
-          "types": [
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "80 Beaver Street, New York"
-      },
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.70526359999999,
-                  "lng": -74.0088519
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.7066125802915,
-                      "lng": -74.00750291970849
-                  },
-                  "southwest": {
-                      "lat": 40.7039146197085,
-                      "lng": -74.01020088029149
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/restaurant-71.png",
-          "id": "ecfe5da9721776d32ebe5d8c7e06da5fd854da0b",
-          "name": "11 Hanover Greek",
-          "opening_hours": {
-              "open_now": true,
-              "weekday_text": []
-          },
-          "photos": [
-              {
-                  "height": 1066,
-                  "html_attributions": [
-                      "<a href=\"https://maps.google.com/maps/contrib/111172608410296664755/photos\">11 Hanover Greek</a>"
-                  ],
-                  "photo_reference": "CmRaAAAA7K131aCDKCWMNBvra1n0UULaTCDKcQ53Fi-2kLywcemVBf9bqM-VHo1FGzgQzHmE2QPoR7P8OBL_4Czi5BGjAIPCbHb7u-sK1AhKbhDnt4IuaWZutSqKBQlCuxAfFxSmEhCtM_0-ndk-3-AxFoa7GSqxGhSwr132KoJuL24-vQfxV4qxkVnrsA",
-                  "width": 1600
-              }
-          ],
-          "place_id": "ChIJORvTZBZawokRbjliKqfPc0w",
-          "price_level": 2,
-          "rating": 4.3,
-          "reference": "CmRRAAAAAOZ2bHOA440EEwQe4vkvLuhf1EqLItAAM7Ox6Py6IcbRoXoj8AJ6A_6CvkrajA0j9Y8aEdqE3Y8qbJ4zWZw0RzvsVETC3F7koqpNEkL80QHfIeypimwHtsvPdhtiISFOEhD0rcaomop5Y9vZI_IuQyD0GhQ9sTzVO4onj04r3wsd0tih8b5ubg",
-          "scope": "GOOGLE",
-          "types": [
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "11 Hanover Square, New York"
-      },
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.7055235,
-                  "lng": -74.0085941
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.70687248029149,
-                      "lng": -74.0072451197085
-                  },
-                  "southwest": {
-                      "lat": 40.70417451970849,
-                      "lng": -74.00994308029151
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/restaurant-71.png",
-          "id": "9766f034a895d2dfe84b0b25bdca9314eb8ac7d4",
-          "name": "Oaxaca Taqueria",
-          "opening_hours": {
-              "open_now": true,
-              "weekday_text": []
-          },
-          "photos": [
-              {
-                  "height": 3024,
-                  "html_attributions": [
-                      "<a href=\"https://maps.google.com/maps/contrib/100192610855103055252/photos\">Michael Karelis</a>"
-                  ],
-                  "photo_reference": "CmRaAAAAYLTbjNxWsDFQKw-HI-krUiNd78lNDlSBd7B0rjJPqvcO8V3MdZahI2JfbP29Q6V7QwgXVV5z1krog07TSbL09GTHxotAu6cstg1XaQAr_ocp8pxvmyobnOmrdsGdSEG6EhAw72TX1pP7tDET7fYayJkeGhQlF8fEdTcE_WtO_YtzmTnCaxZavA",
-                  "width": 4032
-              }
-          ],
-          "place_id": "ChIJ2WzzaBZawokRzWfSxS4FLkU",
-          "rating": 4.3,
-          "reference": "CmRRAAAAM6w3v0iVBMYsf0qaSpEHVc7F2WEhHPCeAOlsb9B6Hlhz31_bMtE7GJK02Oe20LGLUyxgccdqBdNiNcj1Unr8Mc_IjEe-M8u05qIkQvpEfMEtp1uKz5xZnc3wh1E58l9UEhDEvHJcuN5UP-WRD0V8VLlTGhS1A91QpWa_87lQWS6gwItwrNHoSg",
-          "scope": "GOOGLE",
-          "types": [
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "87 Beaver St, New York"
-      },
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.7051603,
-                  "lng": -74.0080959
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.7066442302915,
-                      "lng": -74.0065949197085
-                  },
-                  "southwest": {
-                      "lat": 40.7039462697085,
-                      "lng": -74.00929288029151
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/restaurant-71.png",
-          "id": "f8b29bdeaf22f092b046d2da96944d9161b7bae5",
-          "name": "Dina Rata",
-          "opening_hours": {
-              "open_now": false,
-              "weekday_text": []
-          },
-          "photos": [
-              {
-                  "height": 2268,
-                  "html_attributions": [
-                      "<a href=\"https://maps.google.com/maps/contrib/115175261554531209507/photos\">Suyeon Lee</a>"
-                  ],
-                  "photo_reference": "CmRZAAAATSSRG-LIFwhr3YwcKAXLbGXoQjgqv-uS3rBZpoJsb77F5XPuccPkeBNSm2ZU48xZEIsCZZpbRWHDu0-Ugo0u3vtKZRnLoNJTrC8CdpsrOvKwd0ZzrSJaCgq7ueSgJehwEhDBpq41YwbmG5tl-kxXVstaGhR-sNVjrfm8DP7p2gnYdbTMs2qPMg",
-                  "width": 4032
-              }
-          ],
-          "place_id": "ChIJF1iRbBZawokR17gk5xAzMIg",
-          "price_level": 2,
-          "rating": 4.2,
-          "reference": "CmRSAAAAdrcAgyKgBArMbmJUqO7op68aVikItJDnuUdEvjfEQ-9PXlkm_bTapYQ2Sed-EvbSdK5tCcql56D342o15S9p25ZRXZH8REcxf1DCNjpAeB64CskkxChG5VuUb9lskijtEhB2aD2d13N-21SXtgyePNRFGhSoJe5ymDJt4G4wkVvkM3Op8_hCeg",
-          "scope": "GOOGLE",
-          "types": [
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "75 Wall Street, New York"
-      },
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.7047329,
-                  "lng": -74.00871409999999
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.70615413029149,
-                      "lng": -74.00743361970849
-                  },
-                  "southwest": {
-                      "lat": 40.7034561697085,
-                      "lng": -74.0101315802915
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/restaurant-71.png",
-          "id": "9deadaa381cab317617bc3ba3c535ad049e18b16",
-          "name": "Jackmonkey catering",
-          "place_id": "ChIJ2SUIehZawokRB2wCi912A54",
-          "reference": "CmRSAAAANXS5lZXZ5CFKqgEO_T1C5ufR-aMBA6TUJlhDU_fOiLf2kCu7TqEdbEWYRFxljlkcaH73lglcFkUrtWiG_iFmjuQXnu0RIPPmsZood86sX68KStoIqqQF69NxOMsqGm5oEhChZLHSWDwA2977uZFP0LBJGhTTWYwh4Y_QOSfQm7sTQdaPyXLg0A",
-          "scope": "GOOGLE",
-          "types": [
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "114 Pearl Street, New York"
-      },
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.7051299,
-                  "lng": -74.00812789999999
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.7065333802915,
-                      "lng": -74.00684311970848
-                  },
-                  "southwest": {
-                      "lat": 40.7038354197085,
-                      "lng": -74.0095410802915
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/restaurant-71.png",
-          "id": "b924617ed917e734dd623a875c0530a5dd7baf74",
-          "name": "Wall Street Grill",
-          "opening_hours": {
-              "open_now": false,
-              "weekday_text": []
-          },
-          "place_id": "ChIJJ0daaxZawokRG_CF1AM2878",
-          "reference": "CmRSAAAAsCXpuJvEbREoj1gBCcHIQisn2WEu-hd9tgsfrnoqPnHEAqKo1HLlzDwk_G49jlVnnYEvPwo1cmq0isk00iVxyUt1IFiSghU07hKp7PtzWVIig4Wiu4v0YdiUy_sKIQnTEhDWnJRDq6Gw6_xioTumBZAlGhQvq3xT2A2yl_NDP44mOCc_LVg0HA",
-          "scope": "GOOGLE",
-          "types": [
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "128 Pearl Street, New York"
-      },
-      {
-          "geometry": {
-              "location": {
-                  "lat": 40.7049126,
-                  "lng": -74.0082204
-              },
-              "viewport": {
-                  "northeast": {
-                      "lat": 40.7061748802915,
-                      "lng": -74.0067689697085
-                  },
-                  "southwest": {
-                      "lat": 40.7034769197085,
-                      "lng": -74.00946693029151
-                  }
-              }
-          },
-          "icon": "https://maps.gstatic.com/mapfiles/place_api/icons/restaurant-71.png",
-          "id": "0f70dc01d9560ad798c6355c5bae30640471b2e0",
-          "name": "Andaz Wall Street Beer Garden",
-          "opening_hours": {
-              "open_now": true,
-              "weekday_text": []
-          },
-          "place_id": "ChIJ0eB2chZawokRc7e7zEZ7Lyo",
-          "reference": "CmRRAAAA0N9e_5Eqx5nvWX42WMtcv0CwiiUnzVSfF42LPZWSaDKbTb1wBhvgoOyQ4Hwbkv3D1bjW_snMtIyFt-XxAi_uBiTli4QEvIoRIdvmo2UzhKdGXzOPdOHFw8qXZE2wC0h2EhC9WDU8h1LVq9kBlz_x6hDFGhSP5eViF9ZMzzNeuxX2FTH-khqyDQ",
-          "scope": "GOOGLE",
-          "types": [
-              "bar",
-              "restaurant",
-              "food",
-              "point_of_interest",
-              "establishment"
-          ],
-          "vicinity": "Water Street & Gouverneur Lane, New York"
-      }
-  ],
-  "status": "OK"
-}
